+++ pypolestar/__init__.py
@@ -2,7 +2,7 @@
 
 from importlib.metadata import version
 
-from .polestar import PolestarApi
+from .api import PolestarApi
 
 __version__ = version("pypolestar")
 __all__ = ["PolestarApi"]
